requirements.txt
numpy==1.25.0
Pillow==9.3.0
Pillow==10.2.0
streamlit==1.14.1
tensorflow==2.15.0.post1
